requirements.txt
jupyterlab==1.1.4
ipywidgets==7.6.3
matplotlib==3.3.3
seaborn==0.11.1
librosa==0.8.0
tensorflow==2.5.0
dm-sonnet==2.0.0
dm-tree==0.1.6
tensorflow-datasets==4.3.0
youtube-dl